CreatePlanButton as function
